+++ jpa-criteria/jpa-2-compatibility/src/main/java/javax/persistence/criteria/CommonAbstractCriteria.java
@@ -1,19 +1,12 @@
 /*
- * Copyright 2014 - 2020 Blazebit.
+ * Copyright (c) 2008, 2009, 2011 Oracle, Inc. All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v1.0 and Eclipse Distribution License v. 1.0
+ * which accompanies this distribution.  The Eclipse Public License is available
+ * at http://www.eclipse.org/legal/epl-v10.html and the Eclipse Distribution License
+ * is available at http://www.eclipse.org/org/documents/edl-v10.php.
  */
-
 package javax.persistence.criteria;
 
 /**
@@ -31,9 +24,11 @@ package javax.persistence.criteria;
  * @since Java Persistence 2.1
  */
 public interface CommonAbstractCriteria {
+
     /**
      * Create a subquery of the query.
-     * @param type  the subquery result type
+     *
+     * @param type the subquery result type
      * @return subquery
      */
     <U> Subquery<U> subquery(Class<U> type);
@@ -42,6 +37,7 @@ public interface CommonAbstractCriteria {
      * Return the predicate that corresponds to the where clause
      * restriction(s), or null if no restrictions have been
      * specified.
+     *
      * @return where clause predicate
      */
     Predicate getRestriction();