+++ jpa-criteria/jpa-2-compatibility/src/main/java/javax/persistence/criteria/CriteriaUpdate.java
@@ -1,19 +1,12 @@
 /*
- * Copyright 2014 - 2020 Blazebit.
+ * Copyright (c) 2008, 2009, 2011 Oracle, Inc. All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v1.0 and Eclipse Distribution License v. 1.0
+ * which accompanies this distribution.  The Eclipse Public License is available
+ * at http://www.eclipse.org/legal/epl-v10.html and the Eclipse Distribution License
+ * is available at http://www.eclipse.org/org/documents/edl-v10.php.
  */
-
 package javax.persistence.criteria;
 
 import javax.persistence.metamodel.EntityType;
@@ -34,10 +27,10 @@ import javax.persistence.metamodel.SingularAttribute;
  * A CriteriaUpdate object must have a single root.
  *
  * @param <T> the entity type that is the target of the update
- *
  * @since Java Persistence 2.1
  */
 public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
+
     /**
      * Create and add a query root corresponding to the entity
      * that is the target of the update.
@@ -45,7 +38,6 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * is being updated.
      *
      * @param entityClass the entity class
-     *
      * @return query root corresponding to the given entity
      */
     Root<T> from(Class<T> entityClass);
@@ -57,7 +49,6 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * is being updated.
      *
      * @param entity metamodel entity representing the entity of type X
-     *
      * @return query root corresponding to the given entity
      */
     Root<T> from(EntityType<T> entity);
@@ -73,8 +64,7 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Update the value of the specified attribute.
      *
      * @param attribute attribute to be updated
-     * @param value new value
-     *
+     * @param value     new value
      * @return the modified query
      */
     <Y, X extends Y> CriteriaUpdate<T> set(SingularAttribute<? super T, Y> attribute, X value);
@@ -83,8 +73,7 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Update the value of the specified attribute.
      *
      * @param attribute attribute to be updated
-     * @param value new value
-     *
+     * @param value     new value
      * @return the modified query
      */
     <Y> CriteriaUpdate<T> set(SingularAttribute<? super T, Y> attribute, Expression<? extends Y> value);
@@ -93,8 +82,7 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Update the value of the specified attribute.
      *
      * @param attribute attribute to be updated
-     * @param value new value
-     *
+     * @param value     new value
      * @return the modified query
      */
     <Y, X extends Y> CriteriaUpdate<T> set(Path<Y> attribute, X value);
@@ -103,8 +91,7 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Update the value of the specified attribute.
      *
      * @param attribute attribute to be updated
-     * @param value new value
-     *
+     * @param value     new value
      * @return the modified query
      */
     <Y> CriteriaUpdate<T> set(
@@ -115,8 +102,7 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Update the value of the specified attribute.
      *
      * @param attributeName name of the attribute to be updated
-     * @param value new value
-     *
+     * @param value         new value
      * @return the modified query
      */
     CriteriaUpdate<T> set(String attributeName, Object value);
@@ -127,7 +113,6 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * Replaces the previously added restriction(s), if any.
      *
      * @param restriction a simple or compound boolean expression
-     *
      * @return the modified query
      */
     CriteriaUpdate<T> where(Expression<Boolean> restriction);
@@ -141,7 +126,6 @@ public interface CriteriaUpdate<T> extends CommonAbstractCriteria {
      * restrictions are simply removed.
      *
      * @param restrictions zero or more restriction predicates
-     *
      * @return the modified query
      */
     CriteriaUpdate<T> where(Predicate... restrictions);