+++ CHANGELOG.md
@@ -13,6 +13,7 @@ None
 ### Bug fixes
 
 * Missing flushes for attributes that were changed in pre-update entity view listeners
+* Render enum literal as parameter if compared with attribute having Hibernate non-`EnhancedUserType`
 
 ### Backwards-incompatible changes